Move BarOptions to separate source file
This is now in line with what we do for the other graph types.
