+++ src/main/scala/co/theasi/plotly/SeriesOptions.scala
@@ -9,20 +9,3 @@ case class BoxOptions(
 ) extends SeriesOptions {
   def name(newName: String): BoxOptions = copy(name = Some(newName))
 }
-
-
-case class BarOptions(
-  name: Option[String] = None,
-  marker: MarkerOptions = MarkerOptions()
-) extends SeriesOptions {
-  def name(newName: String): BarOptions = copy(name = Some(newName))
-
-  def marker(newMarker: MarkerOptions): BarOptions =
-    copy(marker = newMarker)
-
-  def updatedMarker(updater: MarkerOptions => MarkerOptions)
-  : BarOptions = {
-    val newMarker = updater(marker)
-    marker(newMarker)
-  }
-}
